Add missing IgnoreHttpsErrors tests
