+++ lib/PuppeteerSharp.Tests/IgnoreHttpsErrorsTests/ResponseSecurityDetailsTests.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
@@ -42,6 +42,14 @@ namespace PuppeteerSharp.Tests.IgnoreHttpsErrorsTests
                 TestUtils.CurateProtocol(response.SecurityDetails.Protocol));
         }
 
+        [PuppeteerTest("ignorehttpserrors.spec.ts", "Response.securityDetails", "should be |null| for non-secure requests")]
+        [SkipBrowserFact(skipFirefox: true)]
+        public async Task ShouldBeNullForNonSecureRequests()
+        {
+            var response = await Page.GoToAsync(TestConstants.EmptyPage);
+            Assert.Null(response.SecurityDetails);
+        }
+
         [PuppeteerTest("ignorehttpserrors.spec.ts", "Response.securityDetails", "Network redirects should report SecurityDetails")]
         [SkipBrowserFact(skipFirefox: true)]
         public async Task NetworkRedirectsShouldReportSecurityDetails()
@@ -67,35 +75,6 @@ namespace PuppeteerSharp.Tests.IgnoreHttpsErrorsTests
             Assert.Equal(
                 TestUtils.CurateProtocol(requestTask.Result.ToString()),
                 TestUtils.CurateProtocol(response.SecurityDetails.Protocol));
-        }
-
-        [PuppeteerTest("ignorehttpserrors.spec.ts", "Response.securityDetails", "should work with request interception")]
-        [SkipBrowserFact(skipFirefox: true)]
-        public async Task ShouldWorkWithRequestInterception()
-        {
-            await Page.SetRequestInterceptionAsync(true);
-            Page.Request += async (_, e) => await e.Request.ContinueAsync();
-            var response = await Page.GoToAsync(TestConstants.EmptyPage);
-            Assert.Equal(HttpStatusCode.OK, response.Status);
-        }
-
-        [PuppeteerTest("ignorehttpserrors.spec.ts", "Response.securityDetails", "should work with mixed content")]
-        [SkipBrowserFact(skipFirefox: true)]
-        public async Task ShouldWorkWithMixedContent()
-        {
-            HttpsServer.SetRoute("/mixedcontent.html", async (context) =>
-            {
-                await context.Response.WriteAsync($"<iframe src='{TestConstants.EmptyPage}'></iframe>");
-            });
-            await Page.GoToAsync(TestConstants.HttpsPrefix + "/mixedcontent.html", new NavigationOptions
-            {
-                WaitUntil = new[] { WaitUntilNavigation.Load }
-            });
-            Assert.Equal(2, Page.Frames.Length);
-            // Make sure blocked iframe has functional execution context
-            // @see https://github.com/GoogleChrome/puppeteer/issues/2709
-            Assert.Equal(3, await Page.MainFrame.EvaluateExpressionAsync<int>("1 + 2"));
-            Assert.Equal(5, await Page.FirstChildFrame().EvaluateExpressionAsync<int>("2 + 3"));
         }
     }
 }
